requirements.txt
fastapi==0.121.3
uvicorn[standard]==0.38.0
python-dotenv==1.2.1
boto3==1.41.1
browser-use==0.9.5
langchain-google-genai==2.0.8
pydantic==2.12.4
pydantic-settings==2.12.0
reportlab==4.2.5
